server: Update existing license grants on license policy changes
