+++ server/polar/license_key/schemas.py
@@ -1,11 +1,14 @@
 from datetime import datetime
-from typing import Any
+from typing import Any, Literal, Self
 
+from dateutil.relativedelta import relativedelta
 from pydantic import UUID4, Field
 
 from polar.benefit.schemas import BenefitID
 from polar.exceptions import ResourceNotFound, Unauthorized
 from polar.kit.schemas import Schema
+from polar.kit.utils import generate_uuid, utc_now
+from polar.models.benefit import BenefitLicenseKeyExpiration
 from polar.models.license_key import LicenseKeyStatus
 
 ###############################################################################
@@ -89,6 +92,60 @@ class LicenseKeyUpdate(Schema):
 
 
 class LicenseKeyCreate(LicenseKeyUpdate):
+    organization_id: UUID4
     user_id: UUID4
     benefit_id: BenefitID
     key: str
+
+    @classmethod
+    def generate_key(cls, prefix: str | None = None) -> str:
+        key = str(generate_uuid()).upper()
+        if prefix is None:
+            return key
+
+        prefix = prefix.strip().upper()
+        return f"{prefix}-{key}"
+
+    @classmethod
+    def generate_expiration_dt(
+        cls, ttl: int, timeframe: Literal["year", "month", "day"]
+    ) -> datetime:
+        now = utc_now()
+        match timeframe:
+            case "year":
+                return now + relativedelta(years=ttl)
+            case "month":
+                return now + relativedelta(months=ttl)
+            case _:
+                return now + relativedelta(days=ttl)
+
+    @classmethod
+    def build(
+        cls,
+        organization_id: UUID4,
+        user_id: UUID4,
+        benefit_id: UUID4,
+        prefix: str | None = None,
+        status: LicenseKeyStatus = LicenseKeyStatus.granted,
+        limit_activations: int | None = None,
+        limit_usage: int | None = None,
+        expires: BenefitLicenseKeyExpiration | None = None,
+    ) -> Self:
+        expires_at = None
+        if expires:
+            ttl = expires.get("ttl", None)
+            timeframe = expires.get("timeframe", None)
+            if ttl and timeframe:
+                expires_at = cls.generate_expiration_dt(ttl, timeframe)
+
+        key = cls.generate_key(prefix=prefix)
+        return cls(
+            organization_id=organization_id,
+            user_id=user_id,
+            benefit_id=benefit_id,
+            key=key,
+            status=status,
+            limit_activations=limit_activations,
+            limit_usage=limit_usage,
+            expires_at=expires_at
+        )
